+++ includes/activitypub/transformer/class-gatherpress.php
@@ -1,19 +1,19 @@
 <?php
 /**
- * ActivityPub Transformer for the plugin Very Simple Event List.
+ * ActivityPub Transformer for the GatherPress event plugin.
  *
- * @package ActivityPub_Event_Bridge
+ * @package Event_Bridge_For_ActivityPub
  * @license AGPL-3.0-or-later
  */
 
-namespace ActivityPub_Event_Bridge\Activitypub\Transformer;
+namespace Event_Bridge_For_ActivityPub\Activitypub\Transformer;
 
 // Exit if accessed directly.
 defined( 'ABSPATH' ) || exit; // @codeCoverageIgnore
 
 use Activitypub\Activity\Extended_Object\Event as Event_Object;
 use Activitypub\Activity\Extended_Object\Place;
-use ActivityPub_Event_Bridge\Activitypub\Transformer\Event;
+use Event_Bridge_For_ActivityPub\Activitypub\Transformer\Event;
 use GatherPress\Core\Event as GatherPress_Event;
 
 /**