+++ includes/activitypub/transformer/class-modern-events-calendar-lite.php
@@ -2,17 +2,17 @@
 /**
  * ActivityPub Tribe Transformer
  *
- * @package ActivityPub_Event_Bridge
+ * @package Event_Bridge_For_ActivityPub
  * @license AGPL-3.0-or-later
  */
 
-namespace ActivityPub_Event_Bridge\Activitypub\Transformer;
+namespace Event_Bridge_For_ActivityPub\Activitypub\Transformer;
 
 // Exit if accessed directly.
 defined( 'ABSPATH' ) || exit; // @codeCoverageIgnore
 
 use Activitypub\Activity\Extended_Object\Place;
-use ActivityPub_Event_Bridge\Activitypub\Transformer\Event;
+use Event_Bridge_For_ActivityPub\Activitypub\Transformer\Event;
 
 use MEC;
 use MEC\Events\Event as MEC_Event;