+++ includes/class-setup.php
@@ -1,32 +1,32 @@
 <?php
 /**
- * Class responsible for initializing ActivityPub Event Bridge.
+ * Class responsible for initializing Event Bridge for ActivityPub.
  *
  * The setup class provides function for checking if this plugin should be activated.
  * It detects supported event plugins and provides all setup hooks and filters.
  *
- * @package ActivityPub_Event_Bridge
+ * @package Event_Bridge_For_ActivityPub
  * @since 1.0.0
  * @license AGPL-3.0-or-later
  */
 
-namespace ActivityPub_Event_Bridge;
+namespace Event_Bridge_For_ActivityPub;
 
 // Exit if accessed directly.
 defined( 'ABSPATH' ) || exit; // @codeCoverageIgnore
 
-use ActivityPub_Event_Bridge\Admin\Event_Plugin_Admin_Notices;
-use ActivityPub_Event_Bridge\Admin\General_Admin_Notices;
-use ActivityPub_Event_Bridge\Admin\Health_Check;
-use ActivityPub_Event_Bridge\Admin\Settings_Page;
-use ActivityPub_Event_Bridge\Plugins\Event_Plugin;
+use Event_Bridge_For_ActivityPub\Admin\Event_Plugin_Admin_Notices;
+use Event_Bridge_For_ActivityPub\Admin\General_Admin_Notices;
+use Event_Bridge_For_ActivityPub\Admin\Health_Check;
+use Event_Bridge_For_ActivityPub\Admin\Settings_Page;
+use Event_Bridge_For_ActivityPub\Integrations\Event_Plugin;
 
 require_once ABSPATH . 'wp-admin/includes/plugin.php';
 
 /**
  * Class Setup.
  *
- * This class is responsible for initializing ActivityPub Event Bridge.
+ * This class is responsible for initializing Event Bridge for ActivityPub.
  *
  * @since 1.0.0
  */
@@ -64,7 +64,7 @@ class Setup {
 			is_plugin_active( 'activitypub/activitypub.php' );
 		// BeforeFirstRelease: decide whether we want to do anything at all when ActivityPub plugin is note active.
 		// if ( ! $this->activitypub_plugin_is_active ) {
-		// deactivate_plugins( ACTIVITYPUB_EVENT_BRIDGE_PLUGIN_FILE );
+		// deactivate_plugins( EVENT_BRIDGE_FOR_ACTIVITYPUB_PLUGIN_FILE );
 		// return;
 		// }.
 		$this->active_event_plugins       = self::detect_active_event_plugins();
@@ -125,13 +125,15 @@ class Setup {
 	 * @var array
 	 */
 	private const EVENT_PLUGIN_CLASSES = array(
-		'\ActivityPub_Event_Bridge\Plugins\Events_Manager',
-		'\ActivityPub_Event_Bridge\Plugins\GatherPress',
-		'\ActivityPub_Event_Bridge\Plugins\The_Events_Calendar',
-		'\ActivityPub_Event_Bridge\Plugins\VS_Event_List',
-		'\ActivityPub_Event_Bridge\Plugins\WP_Event_Manager',
-		'\ActivityPub_Event_Bridge\Plugins\Eventin',
-		'\ActivityPub_Event_Bridge\Plugins\Modern_Events_Calendar_Lite',
+		'\Event_Bridge_For_ActivityPub\Integrations\Events_Manager',
+		'\Event_Bridge_For_ActivityPub\Integrations\GatherPress',
+		'\Event_Bridge_For_ActivityPub\Integrations\The_Events_Calendar',
+		'\Event_Bridge_For_ActivityPub\Integrations\VS_Event_List',
+		'\Event_Bridge_For_ActivityPub\Integrations\WP_Event_Manager',
+		'\Event_Bridge_For_ActivityPub\Integrations\Eventin',
+		'\Event_Bridge_For_ActivityPub\Integrations\Modern_Events_Calendar_Lite',
+		'\Event_Bridge_For_ActivityPub\Integrations\EventPrime',
+		'\Event_Bridge_For_ActivityPub\Integrations\Event_Organiser',
 	);
 
 	/**
@@ -164,14 +166,14 @@ class Setup {
 	 * @return void
 	 */
 	protected function setup_hooks(): void {
-		register_activation_hook( ACTIVITYPUB_EVENT_BRIDGE_PLUGIN_FILE, array( $this, 'activate' ) );
+		register_activation_hook( EVENT_BRIDGE_FOR_ACTIVITYPUB_PLUGIN_FILE, array( $this, 'activate' ) );
 
 		add_action( 'admin_init', array( $this, 'do_admin_notices' ) );
 		add_action( 'admin_init', array( Settings::class, 'register_settings' ) );
 		add_action( 'admin_enqueue_scripts', array( self::class, 'enqueue_styles' ) );
 		add_action( 'admin_menu', array( Settings_Page::class, 'admin_menu' ) );
 		add_filter(
-			'plugin_action_links_' . ACTIVITYPUB_EVENT_BRIDGE_PLUGIN_BASENAME,
+			'plugin_action_links_' . EVENT_BRIDGE_FOR_ACTIVITYPUB_PLUGIN_BASENAME,
 			array( Settings_Page::class, 'settings_link' )
 		);
 
@@ -184,7 +186,7 @@ class Setup {
 		add_action( 'init', array( Event_Sources::class, 'register_taxonomy' ) );
 
 		// Check if the minimum required version of the ActivityPub plugin is installed.
-		if ( ! version_compare( $this->activitypub_plugin_version, ACTIVITYPUB_EVENT_BRIDGE_ACTIVITYPUB_PLUGIN_MIN_VERSION ) ) {
+		if ( ! version_compare( $this->activitypub_plugin_version, EVENT_BRIDGE_FOR_ACTIVITYPUB_ACTIVITYPUB_PLUGIN_MIN_VERSION ) ) {
 			return;
 		}
 
@@ -199,24 +201,24 @@ class Setup {
 	 * @return void
 	 */
 	public static function enqueue_styles( $hook_suffix ): void {
-		if ( false !== strpos( $hook_suffix, 'activitypub-event-bridge' ) ) {
+		if ( false !== strpos( $hook_suffix, 'event-bridge-for-activitypub' ) ) {
 			wp_enqueue_style(
-				'activitypub-event-bridge-admin-styles',
+				'event-bridge-for-activitypub-admin-styles',
 				plugins_url(
-					'assets/css/activitypub-event-bridge-admin.css',
-					ACTIVITYPUB_EVENT_BRIDGE_PLUGIN_FILE
+					'assets/css/event-bridge-for-activitypub-admin.css',
+					EVENT_BRIDGE_FOR_ACTIVITYPUB_PLUGIN_FILE
 				),
 				array(),
-				ACTIVITYPUB_EVENT_BRIDGE_PLUGIN_VERSION
+				EVENT_BRIDGE_FOR_ACTIVITYPUB_PLUGIN_VERSION
 			);
 			wp_enqueue_script(
-				'activitypub-event-bridge-admin-script',
+				'event-bridge-for-activitypub-admin-script',
 				plugins_url(
-					'assets/js/activitypub-event-bridge-admin.js',
-					ACTIVITYPUB_EVENT_BRIDGE_PLUGIN_FILE
+					'assets/js/event-bridge-for-activitypub-admin.js',
+					EVENT_BRIDGE_FOR_ACTIVITYPUB_PLUGIN_FILE
 				),
 				array( 'jquery' ),
-				ACTIVITYPUB_EVENT_BRIDGE_PLUGIN_VERSION,
+				EVENT_BRIDGE_FOR_ACTIVITYPUB_PLUGIN_VERSION,
 				false
 			);
 		}
@@ -232,15 +234,15 @@ class Setup {
 		// Check if any general admin notices are needed and add actions to insert the needed admin notices.
 		if ( ! $this->activitypub_plugin_is_active ) {
 			// The ActivityPub plugin is not active.
-			add_action( 'admin_notices', array( 'ActivityPub_Event_Bridge\Admin\General_Admin_Notices', 'activitypub_plugin_not_enabled' ), 10, 1 );
+			add_action( 'admin_notices', array( 'Event_Bridge_For_ActivityPub\Admin\General_Admin_Notices', 'activitypub_plugin_not_enabled' ), 10, 1 );
 		}
-		if ( ! version_compare( $this->activitypub_plugin_version, ACTIVITYPUB_EVENT_BRIDGE_ACTIVITYPUB_PLUGIN_MIN_VERSION ) ) {
+		if ( ! version_compare( $this->activitypub_plugin_version, EVENT_BRIDGE_FOR_ACTIVITYPUB_ACTIVITYPUB_PLUGIN_MIN_VERSION ) ) {
 			// The ActivityPub plugin is too old.
-			add_action( 'admin_notices', array( 'ActivityPub_Event_Bridge\Admin\General_Admin_Notices', 'activitypub_plugin_version_too_old' ), 10, 1 );
+			add_action( 'admin_notices', array( 'Event_Bridge_For_ActivityPub\Admin\General_Admin_Notices', 'activitypub_plugin_version_too_old' ), 10, 1 );
 		}
 		if ( empty( $this->active_event_plugins ) ) {
 			// No supported Event Plugin is active.
-			add_action( 'admin_notices', array( 'ActivityPub_Event_Bridge\Admin\General_Admin_Notices', 'no_supported_event_plugin_active' ), 10, 1 );
+			add_action( 'admin_notices', array( 'Event_Bridge_For_ActivityPub\Admin\General_Admin_Notices', 'no_supported_event_plugin_active' ), 10, 1 );
 		}
 	}
 
@@ -253,7 +255,7 @@ class Setup {
 	 *
 	 * @return \Activitypub\Transformer\Base|null
 	 */
-	public function register_activitypub_event_transformer( $transformer, $wp_object, $object_class ): \Activitypub\Transformer\Base {
+	public function register_activitypub_event_transformer( $transformer, $wp_object, $object_class ): ?\Activitypub\Transformer\Base {
 		// If the current WordPress object is not a post (e.g., a WP_Comment), don't change the transformer.
 		if ( 'WP_Post' !== $object_class ) {
 			return $transformer;
@@ -293,9 +295,9 @@ class Setup {
 	}
 
 	/**
-	 * Activates the ActivityPub Event Bridge plugin.
+	 * Activates the Event Bridge for ActivityPub plugin.
 	 *
-	 * This method handles the activation of the ActivityPub Event Bridge plugin.
+	 * This method handles the activation of the Event Bridge for ActivityPub plugin.
 	 *
 	 * @since 1.0.0
 	 * @see register_activation_hook()
@@ -304,7 +306,7 @@ class Setup {
 	public function activate(): void {
 		// Don't allow plugin activation, when the ActivityPub plugin is not activated yet.
 		if ( ! $this->activitypub_plugin_is_active ) {
-			deactivate_plugins( plugin_basename( ACTIVITYPUB_EVENT_BRIDGE_PLUGIN_FILE ) );
+			deactivate_plugins( plugin_basename( EVENT_BRIDGE_FOR_ACTIVITYPUB_PLUGIN_FILE ) );
 			$notice = General_Admin_Notices::get_admin_notice_activitypub_plugin_not_enabled();
 			wp_die(
 				wp_kses( $notice, General_Admin_Notices::ALLOWED_HTML ),
@@ -314,7 +316,7 @@ class Setup {
 		}
 
 		if ( empty( $this->active_event_plugins ) ) {
-			deactivate_plugins( plugin_basename( ACTIVITYPUB_EVENT_BRIDGE_PLUGIN_FILE ) );
+			deactivate_plugins( plugin_basename( EVENT_BRIDGE_FOR_ACTIVITYPUB_PLUGIN_FILE ) );
 			$notice = General_Admin_Notices::get_admin_notice_no_supported_event_plugin_active();
 			wp_die(
 				wp_kses( $notice, General_Admin_Notices::ALLOWED_HTML ),