+++ includes/integrations/class-eventin.php
@@ -1,18 +1,15 @@
 <?php
 /**
- * VS Events LIst.
+ * The Events Calendar.
  *
- * Defines all the necessary meta information for the WordPress event plugin
- * "Very Simple Events List".
+ * Defines all the necessary meta information for the events calendar.
  *
- * @link    https://de.wordpress.org/plugins/very-simple-event-list/
- * @package ActivityPub_Event_Bridge
+ * @link    https://wordpress.org/plugins/the-events-calendar/
+ * @package Event_Bridge_For_ActivityPub
  * @since   1.0.0
  */
 
-namespace ActivityPub_Event_Bridge\Plugins;
-
-use ActivityPub_Event_Bridge\Event_Plugins;
+namespace Event_Bridge_For_ActivityPub\Integrations;
 
 // Exit if accessed directly.
 defined( 'ABSPATH' ) || exit; // @codeCoverageIgnore
@@ -24,14 +21,14 @@ defined( 'ABSPATH' ) || exit; // @codeCoverageIgnore
  *
  * @since 1.0.0
  */
-final class VS_Event_List extends Event_Plugin {
+final class Eventin extends Event_plugin {
 	/**
 	 * Returns the full plugin file.
 	 *
 	 * @return string
 	 */
 	public static function get_plugin_file(): string {
-		return 'very-simple-event-list/vsel.php';
+		return 'wp-event-solution/eventin.php';
 	}
 
 	/**
@@ -40,25 +37,16 @@ final class VS_Event_List extends Event_Plugin {
 	 * @return string
 	 */
 	public static function get_post_type(): string {
-		return 'event';
+		return 'etn';
 	}
 
 	/**
 	 * Returns the IDs of the admin pages of the plugin.
 	 *
-	 * @return array The settings page urls.
+	 * @return array The settings page url.
 	 */
 	public static function get_settings_pages(): array {
-		return array( 'settings_page_vsel' );
-	}
-
-	/**
-	 * Returns the ActivityPub transformer class.
-	 *
-	 * @return string
-	 */
-	public static function get_activitypub_transformer_class_name(): string {
-		return 'VS_Event';
+		return array( 'eventin' ); // Base always is wp-admin/admin.php?page=eventin.
 	}
 
 	/**
@@ -67,6 +55,6 @@ final class VS_Event_List extends Event_Plugin {
 	 * @return string
 	 */
 	public static function get_event_category_taxonomy(): string {
-		return 'event_cat';
+		return 'etn_category';
 	}
 }