+++ includes/integrations/class-vs-event-list.php
@@ -1,15 +1,18 @@
 <?php
 /**
- * Modern Events Calendar (Lite)
+ * VS Events LIst.
  *
- * Defines all the necessary meta information for the Modern Events Calendar (Lite).
+ * Defines all the necessary meta information for the WordPress event plugin
+ * "Very Simple Events List".
  *
- * @link    https://webnus.net/modern-events-calendar/
- * @package ActivityPub_Event_Bridge
+ * @link    https://de.wordpress.org/plugins/very-simple-event-list/
+ * @package Event_Bridge_For_ActivityPub
  * @since   1.0.0
  */
 
-namespace ActivityPub_Event_Bridge\Plugins;
+namespace Event_Bridge_For_ActivityPub\Integrations;
+
+use Event_Bridge_For_ActivityPub\Event_Plugins;
 
 // Exit if accessed directly.
 defined( 'ABSPATH' ) || exit; // @codeCoverageIgnore
@@ -21,14 +24,14 @@ defined( 'ABSPATH' ) || exit; // @codeCoverageIgnore
  *
  * @since 1.0.0
  */
-final class Modern_Events_Calendar_Lite extends Event_plugin {
+final class VS_Event_List extends Event_Plugin {
 	/**
 	 * Returns the full plugin file.
 	 *
 	 * @return string
 	 */
 	public static function get_plugin_file(): string {
-		return 'modern-events-calendar-lite/modern-events-calendar-lite.php';
+		return 'very-simple-event-list/vsel.php';
 	}
 
 	/**
@@ -37,8 +40,7 @@ final class Modern_Events_Calendar_Lite extends Event_plugin {
 	 * @return string
 	 */
 	public static function get_post_type(): string {
-		// See MEC_feature_events->get_main_post_type().
-		return apply_filters( 'mec_post_type_name', 'mec-events' ); // phpcs:ignore
+		return 'event';
 	}
 
 	/**
@@ -47,7 +49,16 @@ final class Modern_Events_Calendar_Lite extends Event_plugin {
 	 * @return array The settings page urls.
 	 */
 	public static function get_settings_pages(): array {
-		return array( 'MEC-settings', 'MEC-support', 'MEC-ix', 'MEC-wizard', 'MEC-addons', 'mec-intro' );
+		return array( 'settings_page_vsel' );
+	}
+
+	/**
+	 * Returns the ActivityPub transformer class.
+	 *
+	 * @return string
+	 */
+	public static function get_activitypub_transformer_class_name(): string {
+		return 'VS_Event';
 	}
 
 	/**
@@ -56,6 +67,6 @@ final class Modern_Events_Calendar_Lite extends Event_plugin {
 	 * @return string
 	 */
 	public static function get_event_category_taxonomy(): string {
-		return 'mec_category';
+		return 'event_cat';
 	}
 }