+++ CHANGELOG.md
@@ -5,8 +5,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.0] - 2024-10-20
+## Unreleased
 
 ### Added
 
-* Initial version tag.
+* Integration for Event Organiser
+* Integration for EventPrime – Events Calendar, Bookings and Tickets
+
+### Fixed
+
+* Fixed that transformer hook function might also return `null`.
+
+## [0.2.1] - 2024-11-16
+
+### Added
+
+* Initial release on WordPress.org
+
+### Fixed
+
+* Applied some WordPress best practices
+
+## [0.2.0] - 2024-10-20
+
+### Added
+
+* Initial submission to WordPress.org